refactor: 패키지 구조 변경 - user-service 하위 패키지에 auth와 user를 설정 - 모듈 분리 통합을 고려해서 auth와 user로 설정했음
