+++ user-service/src/main/java/com/sparta/userservice/UserServiceApplication.java
@@ -2,8 +2,10 @@ package com.sparta.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableFeignClients
 @EnableJpaAuditing
 @SpringBootApplication
 public class UserServiceApplication {
