--- user-service/src/main/java/com/sparta/userservice/domain/repository/CompanyManagerRepository.java
@@ -1,8 +1,0 @@
-package com.sparta.userservice.domain.repository;
-
-import com.sparta.userservice.domain.model.CompanyManager;
-
-public interface CompanyManagerRepository {
-
-	CompanyManager save(CompanyManager companyManager);
-}
