--- user-service/src/main/java/com/sparta/userservice/domain/repository/HubManagerRepository.java
@@ -1,8 +1,0 @@
-package com.sparta.userservice.domain.repository;
-
-import com.sparta.userservice.domain.model.HubManager;
-
-public interface HubManagerRepository {
-
-	HubManager save(HubManager hubManager);
-}
