+++ user-service/src/main/java/com/sparta/userservice/user/domain/repository/CompanyManagerRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.userservice.user.domain.repository;
+
+import com.sparta.userservice.user.domain.model.CompanyManager;
+
+public interface CompanyManagerRepository {
+
+	CompanyManager save(CompanyManager companyManager);
+}
