+++ user-service/src/main/java/com/sparta/userservice/user/domain/repository/HubManagerRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.userservice.user.domain.repository;
+
+import com.sparta.userservice.user.domain.model.HubManager;
+
+public interface HubManagerRepository {
+
+	HubManager save(HubManager hubManager);
+}
