+++ user-service/src/main/java/com/sparta/userservice/user/infra/repository/CompanyManagerJpaRepository.java
@@ -1,10 +1,10 @@
-package com.sparta.userservice.infra.repository;
+package com.sparta.userservice.user.infra.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sparta.userservice.domain.model.CompanyManager;
+import com.sparta.userservice.user.domain.model.CompanyManager;
 
 public interface CompanyManagerJpaRepository extends JpaRepository<CompanyManager, UUID> {
 }
