+++ user-service/src/main/java/com/sparta/userservice/user/infra/repository/HubManagerJpaRepository.java
@@ -1,10 +1,10 @@
-package com.sparta.userservice.infra.repository;
+package com.sparta.userservice.user.infra.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sparta.userservice.domain.model.HubManager;
+import com.sparta.userservice.user.domain.model.HubManager;
 
 public interface HubManagerJpaRepository extends JpaRepository<HubManager, UUID> {
 }
